apps/qt_respeaker_app/requirements.txt
PyAudio
click
numpy
pixel-ring
pyusb
